pyproject.toml
[tool.poetry]
name = "gamestone"
version = "0.1.0"
description = ""
authors = ["Riley Mathews"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
gunicorn = "^20.1.0"
dj-database-url = "^1.2.0"
psycopg2 = "^2.9.5"
whitenoise = "^6.3.0"
requests = "^2.28.1"
authlib = "^1.2.0"
drf-spectacular = "^0.25.1"
django-widget-tweaks = "^1.4.12"
django-model-utils = "^4.3.1"
django-debug-toolbar = "^3.8.1"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^22.12.0"
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
